+++ cypress/support/mocks/fhir-server-mocks.ts
@@ -1,3 +1,0 @@
-export const searchFHIRServer = (titleSearchTerm, fixtureFile) => {
-  cy.intercept(`**title:contains=${titleSearchTerm}**`, {fixture: fixtureFile}).as('searchFHIRServer');
-};