+++ src/app/lib/widgets/lfb-spinner/lfb-spinner.component.spec.ts
@@ -8,7 +8,7 @@ describe('LfbSpinnerComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [ LfbSpinnerComponent ]
+      imports: [ LfbSpinnerComponent ]
     })
     .compileComponents();
   });