+++ src/app/services/form.service.ts
@@ -6,11 +6,12 @@ import {IDType, ITreeNode} from '@bugsplat/angular-tree-component/lib/defs/api';
 import {TreeModel, TreeNode} from '@bugsplat/angular-tree-component';
 import fhir from 'fhir/r4';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
-import {MessageDlgComponent, MessageType} from '../lib/widgets/message-dlg/message-dlg.component';
+import {MessageDlgComponent, MessageDlgOptions, MessageType} from '../lib/widgets/message-dlg/message-dlg.component';
 import {Observable, Subject} from 'rxjs';
 import {HttpClient} from '@angular/common/http';
 import jsonTraverse from 'traverse';
 import {JsonPointer} from 'json-ptr';
+import {JSONPath} from 'jsonpath-plus';
 import {loadLForms, getSupportedLFormsVersions} from 'lforms-loader';
 
 // Configuration files
@@ -25,13 +26,16 @@ import ngxFlSchema from '../../assets/ngx-fl.schema.json5';
 // @ts-ignore
 import flLayout from '../../assets/fl-fields-layout.json5';
 // @ts-ignore
-import itemEditorSchema from '../../assets/item-editor.schema.json5';
+import vsLayout from '../../assets/value-set-fields-layout.json5';
+// @ts-ignore
+import ngxVSSchema from '../../assets/ngx-vs.schema.json5';
+
 import {GuidingStep, Util, FHIR_VERSIONS, FHIR_VERSION_TYPE} from '../lib/util';
 import {FetchService} from './fetch.service';
 import {TerminologyServerComponent} from '../lib/widgets/terminology-server/terminology-server.component';
 import {ExtensionsService} from './extensions.service';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
-import { toUpper } from 'cypress/types/lodash';
+import { EXTENSION_URL_VARIABLE, EXTENSION_URL_INITIAL_EXPRESSION, EXTENSION_URL_CALCULATED_EXPRESSION, EXTENSION_URL_CUSTOM_VARIABLE_TYPE } from '../lib/constants/constants';
 
 declare var LForms: any;
 
@@ -64,25 +68,23 @@ export class FormService {
   static readonly TREE_NODE_ID = "__$treeNodeId";
   _validationStatusChanged$: Subject<void> = new Subject<void>();
 
-  static INITIAL_EXPRESSION = 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression';
-  static CALCULATED_EXPRESSION = 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-calculatedExpression';
-  static VARIABLE = 'http://hl7.org/fhir/StructureDefinition/variable';
-  static CUSTOM_EXT_VARIABLE_TYPE = 'http://lhcforms.nlm.nih.gov/fhirExt/expression-editor-variable-type';
-
   private _loading = false;
   _guidingStep$: Subject<GuidingStep> = new Subject<GuidingStep>();
   _formReset$: Subject<void> = new Subject<void>();
   _formChanged$: Subject<SimpleChange> = new Subject<SimpleChange>();
   _advPanelState = {
     formLevel: true,
-    itemLevel: true
+    itemLevel: true,
+    valueSet: true,
+    binary: true
   }
 
   localStorageError: Error = null;
   treeModel: TreeModel;
   itemSchema: any = {properties: {}};
   flSchema: any = {properties: {}};
-  private _itemEditorSchema: any = {properties: {}};
+  valueSetSchema: any = {properties: {}};
+  binarySchema: any = {properties: {}};
 
   snomedUser = false;
   _lformsVersion = '';
@@ -136,18 +138,30 @@ export class FormService {
   };
 
   constructor(private modalService: NgbModal, private http: HttpClient, private liveAnnouncer: LiveAnnouncer ) {
-    [{schema: ngxItemSchema as any, layout: itemLayout}, {schema: ngxFlSchema as any, layout: flLayout}].forEach((obj) => {
+    const binarySchema = JSON.parse(JSON.stringify(fhirSchemaDefinitions.definitions.Binary));
+
+    [
+      {schema: ngxItemSchema as any, layout: itemLayout},
+      {schema: ngxFlSchema as any, layout: flLayout},
+      {schema: ngxVSSchema as any, layout: vsLayout},
+      {schema: binarySchema as any, layout: {}} // TODO - Place holder for Binary layout
+    ].forEach((obj) => {
       if(!obj.schema.definitions) {
         obj.schema.definitions = {};
       }
-      obj.schema.definitions = fhirSchemaDefinitions.definitions as any;
+      obj.schema.definitions = JSON.parse(JSON.stringify(fhirSchemaDefinitions.definitions));
       obj.schema.formLayout = obj.layout.formLayout;
       this.overrideSchemaWidgetFromLayout(obj.schema, obj.layout);
       this.overrideFieldLabelsFromLayout(obj.schema, obj.layout);
     });
     this.itemSchema = ngxItemSchema;
     this.flSchema = ngxFlSchema;
-    this._itemEditorSchema = itemEditorSchema;
+    this.valueSetSchema = ngxVSSchema;
+    delete this.valueSetSchema.definitions.ValueSet;
+    delete this.valueSetSchema.definitions.ResourceList;
+    delete this.valueSetSchema.properties.contained;
+    this.binarySchema = binarySchema;
+    delete this.binarySchema.definitions.Binary;
 
 
     // Load lforms.
@@ -160,6 +174,23 @@ export class FormService {
       FormService._lformsLoaded$.error(error);
     });
 
+  }
+
+
+  /**
+   * Get resource schema based on the resource type.
+   *
+   * @param resourceType
+   */
+  getResourceSchema(resourceType: 'ValueSet' | 'Binary') {
+    let ret = null;
+    if(resourceType === 'Binary') {
+      ret = this.binarySchema;
+    }
+    else if(resourceType === 'ValueSet') {
+      ret = this.valueSetSchema;
+    }
+    return ret;
   }
 
   /**
@@ -178,12 +209,15 @@ export class FormService {
     Object.keys(widgetsMap).forEach((widgetType) => {
       const widgetInfo = widgets[widgetType];
       if(widgetInfo) {
-        const fieldPtrs: string[] = widgetsMap[widgetType];
-        fieldPtrs?.forEach((ptr) => {
-          const fieldSchema: any = JsonPointer.get(schema, ptr);
-          if(fieldSchema) {
-            fieldSchema.widget = widgetInfo;
-          }
+        const fieldJsonPaths: string[] = widgetsMap[widgetType];
+        fieldJsonPaths?.forEach((path) => {
+          const pointers = JSONPath({path, json: schema, resultType: 'pointer'});
+          pointers.forEach((ptr: string) => {
+            const fieldSchema: any = JsonPointer.get(schema, ptr);
+            if(fieldSchema) {
+              fieldSchema.widget = widgetInfo;
+            }
+          });
         });
       }
     });
@@ -200,18 +234,17 @@ export class FormService {
       return;
     }
 
-    Object.entries(overridePropertyLabels).forEach(([ptr, title]) => {
-      const fieldSchema: any = JsonPointer.get(schema, ptr);
-      if(fieldSchema) {
+    Object.entries(overridePropertyLabels).forEach(([path, title]) => {
+      const jPointers = JSONPath({path, json: schema, resultType: 'pointer'});
+      jPointers.forEach((ptr: string) => {
+        const fieldSchema: any = JsonPointer.get(schema, ptr);
         if(fieldSchema) {
-          fieldSchema.title = title;
+          if(fieldSchema) {
+            fieldSchema.title = title;
+          }
         }
-      }
+      });
     });
-  }
-
-  public get itemEditorSchema() {
-    return this._itemEditorSchema;
   }
 
   /**
@@ -326,6 +359,34 @@ export class FormService {
    */
   get itemLevel(): boolean {
     return this._advPanelState.itemLevel;
+  }
+
+  /**
+   * Setter for item level advanced panel state
+   */
+  set valueSetResource(collapse: boolean) {
+    this._advPanelState.valueSet = collapse;
+  }
+
+  /**
+   * Getter for item level advanced panel state
+   */
+  get valueSetResource(): boolean {
+    return this._advPanelState.valueSet;
+  }
+
+  /**
+   * Setter for item level advanced panel state
+   */
+  set binaryResource(collapse: boolean) {
+    this._advPanelState.binary = collapse;
+  }
+
+  /**
+   * Getter for item level advanced panel state
+   */
+  get binaryResource(): boolean {
+    return this._advPanelState.binary;
   }
 
   /**
@@ -644,7 +705,7 @@ export class FormService {
    * @returns - returns the array of extensions excluding the extensions of type Variable.
    */
   removeVariablesExtensions(extensions: any): any {
-    return extensions.filter(ext => (ext.url !== FormService.VARIABLE));
+    return extensions.filter(ext => (ext.url !== EXTENSION_URL_VARIABLE));
   }
 
   /**
@@ -653,7 +714,7 @@ export class FormService {
    * @returns - returns the array of extensions excluding the extensions of type Initial Expression or Calculated Expression.
    */
   removeExpressionsExtensions(extensions: any): any {
-    return extensions.filter(ext => (ext.url !== FormService.INITIAL_EXPRESSION && ext.url !== FormService.CALCULATED_EXPRESSION));
+    return extensions.filter(ext => (ext.url !== EXTENSION_URL_INITIAL_EXPRESSION && ext.url !== EXTENSION_URL_CALCULATED_EXPRESSION));
   }
 
   /**
@@ -662,7 +723,7 @@ export class FormService {
    * @returns - returns the array of extensions that is of type Variable.
    */
   filterVariableExtensions(extensions: any): any {
-    return extensions.filter(ext => (ext.url === FormService.VARIABLE));
+    return extensions.filter(ext => (ext.url === EXTENSION_URL_VARIABLE));
   }
 
   /**
@@ -671,7 +732,7 @@ export class FormService {
    * @returns - expression url.
    */
   getUrlByType(expressionType: string): string {
-    return (expressionType === "__$initialExpression") ? FormService.INITIAL_EXPRESSION : FormService.CALCULATED_EXPRESSION;
+    return (expressionType === "__$initialExpression") ? EXTENSION_URL_INITIAL_EXPRESSION : EXTENSION_URL_CALCULATED_EXPRESSION;
   }
 
   /**
@@ -685,10 +746,10 @@ export class FormService {
     let expression = null;
 
     for (const ext of extensions) {
-      if (ext.url === FormService.CALCULATED_EXPRESSION) {
+      if (ext.url === EXTENSION_URL_CALCULATED_EXPRESSION) {
         ext.url = this.getUrlByType(expressionType);
         return ext;
-      } else if (ext.url === FormService.INITIAL_EXPRESSION && !expression) {
+      } else if (ext.url === EXTENSION_URL_INITIAL_EXPRESSION && !expression) {
         expression = ext;
         expression.url = this.getUrlByType(expressionType);
       }
@@ -703,7 +764,7 @@ export class FormService {
    * @returns - True if the Initial Expression is found and false otherwise.
    */
   hasInitialComputeValueExpression(extensions: any): boolean {
-    return extensions.some(ext => (ext.url === FormService.INITIAL_EXPRESSION));
+    return extensions.some(ext => (ext.url === EXTENSION_URL_INITIAL_EXPRESSION));
   }
 
   /**
@@ -712,7 +773,7 @@ export class FormService {
    * @returns - True if the Calculated Expression is found and false otherwise.
    */
   hasContinuouslyComputeValueExpression(extensions: any): boolean {
-    return extensions.some(ext => (ext.url === FormService.CALCULATED_EXPRESSION));
+    return extensions.some(ext => (ext.url === EXTENSION_URL_CALCULATED_EXPRESSION));
   }
 
   /**
@@ -963,12 +1024,13 @@ export class FormService {
    * @param message - Message to display
    * @param type - INFO | WARNING | DANGER
    */
-  showMessage(title: string, message: string, type: MessageType = MessageType.INFO) {
-
+  showMessage(title: string, message: string, type: MessageType = MessageType.INFO, options?: MessageDlgOptions) {
+    options = options || {};
+    options.title = title || options.title;
+    options.message = message || options.message;
+    options.type = type || options.type;
     const modalRef = this.modalService.open(MessageDlgComponent, {scrollable: true});
-    modalRef.componentInstance.title = title;
-    modalRef.componentInstance.message = message;
-    modalRef.componentInstance.type = type;
+    modalRef.componentInstance.options = options;
   }
 
 
@@ -1044,7 +1106,7 @@ export class FormService {
     // Remove any meta.tag.code generated by LForms.
     if(questionnaire.meta?.tag) {
       questionnaire.meta.tag = questionnaire.meta.tag.filter((tag) => {
-        return ! /^\s*lformsVersion\s*:/i.test(tag.code);
+        return ! /^\s*(lformsVersion\s*:)|(lhc-qnvconv-)/i.test(tag.code);
       });
       if(questionnaire.meta.tag.length === 0) {
         delete questionnaire.meta.tag;
@@ -1127,11 +1189,11 @@ export class FormService {
     if(this._storageAvailable('localStorage')) {
       if(value) {
         if(key !== 'state' && value) {
-          // Non state are objects
+          // Non-state are objects
           localStorage.setItem(key, JSON.stringify(value));
         }
         else {
-          // State is string type.
+          // State is a string type.
           localStorage.setItem(key, value);
         }
       }