+++ src/app/services/shared-object.service.ts
@@ -24,7 +24,13 @@ export class SharedObjectService {
     return new AppJsonPipe().transform(item);
     // return JSON.stringify(item, null, 2);
   }));
-  constructor() {}
+
+  _questionnaire: fhir.Questionnaire;
+  constructor() {
+    this.questionnaire$.subscribe(questionnaire => {
+      this._questionnaire = questionnaire;
+    });
+  }
 
 
   /**
@@ -46,13 +52,27 @@ export class SharedObjectService {
     }
   }
 
-
+  /**
+   * Set questionnaire and share it with listeners.
+   *
+   * @param questionnaire - Questionnaire object to share.
+   */
   set questionnaire(questionnaire: fhir.Questionnaire) {
     this.questionnaireSource$.next(questionnaire);
   }
 
+  /**
+   * Get the observable to listen to the changes in questionnaire.
+   */
   get questionnaire$(): Observable<fhir.Questionnaire> {
     return this.questionnaireSource$.asObservable();
+  }
+
+  /**
+   *
+   */
+  get questionnaire(): fhir.Questionnaire {
+    return this._questionnaire;
   }
 
   set currentItem (item: fhir.QuestionnaireItem) {